fix controller bug

+++ init.sql
@@ -1,7 +1,7 @@
 CREATE SCHEMA IF NOT EXISTS postal;
 
 CREATE TABLE IF NOT EXISTS postal.postcodes (
-    id INTEGER PRIMARY KEY,
+    id LONG PRIMARY KEY,
     postcode VARCHAR(10) NOT NULL,
     latitude FLOAT8,
     longitude FLOAT8
